python/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clr-loader==0.2.6
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.6
pycparser==2.21
pyserial==3.5
pythonnet==3.0.3
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
